backend/requirements.txt
transformers>=4.30.0
tokenizers>=0.13.3
nibabel>=5.1.0       
opencv-python-headless>=4.8.0  
pillow>=10.0.0

streamlit>=1.28.0

numpy<2.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0

flask
flask-cors
numpy
opencv-python-headless
nibabel
transformers==4.36.2
pillow